dev: improve workflow cache

+++ tools/docs/make/touch.py
@@ -1,4 +1,4 @@
-"""Touch DEV_BUILD_FLAG."""
+"""Touch files."""
 
 import argparse
 from pathlib import Path
@@ -7,20 +7,22 @@ from pathlib import Path
 def main() -> None:
     """
     The main method.
-    Touch DEV_BUILD_FLAG.
+    Touch files.
     """
 
     parser = argparse.ArgumentParser(
         prog="touch.py",
-        description="Touch DEV_BUILD_FLAG",
+        description="Touch files.",
     )
     parser.add_argument("-f", "--flag", type=str, required=True)
+    parser.add_argument("--check", action="store_true")
     args = parser.parse_args()
 
     flag = Path(args.flag)
 
-    with open(flag, "w", encoding="utf8") as file_handler:
-        file_handler.write("")
+    if not args.check or flag.exists():
+        with open(flag, "w", encoding="utf8") as file_handler:
+            file_handler.write("")
 
 
 main()
